Refactored code from the view to put the logic in the viewmodel for the thread reply module
